# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kinopoisk API",
        "description": "API для управления жанрами кинопоиска",
        "contact": {
            "name": "Kinopoisk API Support",
            "email": "admin@kinopoisk.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Список всех жанров",
                "description": "Возвращает пагинированный список всех жанров",
                "operationId": "6bdd3e3eae144bc764782732920e8144",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Genre"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Создать новый жанр",
                "description": "Создаёт новый жанр в системе",
                "operationId": "4af0d8bb8edbea0b910e67dffb790842",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название жанра",
                                        "type": "string",
                                        "example": "Фантастика"
                                    },
                                    "slug": {
                                        "description": "URL-slug жанра",
                                        "type": "string",
                                        "example": "fantastika"
                                    },
                                    "description": {
                                        "description": "Описание жанра",
                                        "type": "string",
                                        "example": "Жанр научной фантастики"
                                    },
                                    "is_active": {
                                        "description": "Активен ли жанр",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Жанр успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/genres/{genre}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Получить жанр по ID",
                "description": "Возвращает детальную информацию о жанре по его ID",
                "operationId": "aa38cf60cd91e4e24afbcc741d38e13e",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "description": "ID жанра",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Жанр не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Genres"
                ],
                "summary": "Обновить жанр",
                "description": "Обновляет данные существующего жанра",
                "operationId": "228730fe2de539696c2e83e6461d0086",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "description": "ID жанра",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название жанра",
                                        "type": "string",
                                        "example": "Научная фантастика"
                                    },
                                    "slug": {
                                        "description": "URL-slug жанра",
                                        "type": "string",
                                        "example": "nauchnaya-fantastika"
                                    },
                                    "description": {
                                        "description": "Описание жанра",
                                        "type": "string",
                                        "example": "Обновлённое описание"
                                    },
                                    "is_active": {
                                        "description": "Активен ли жанр",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Жанр успешно обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Жанр не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Удалить жанр",
                "description": "Удаляет жанр из системы вместе со связанными фильмами",
                "operationId": "0d66a3babf29b568beb7c3e1fb4baf47",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "description": "ID жанра",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Жанр успешно удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Жанр был удален!"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Жанр не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Genre": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Фантастика"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "fantastika"
                    },
                    "description": {
                        "type": "string",
                        "example": "Жанр научной фантастики"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "18-08-2026 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "18-08-2026 12:00:00"
                    },
                    "movies": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "title": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "JWT токен авторизации. Формат: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}